Refactored OutlineRendererFeature and improved AnisotropicKuwahara integration and performance
- Added new enum `OutlineSource` with DepthOnly, NormalOnly, and DepthAndNormal modes.
- Implemented dynamic shader keyword control for outline source selection.
- Added automatic shader loading for OutlineRendererFeature and other custom RenderFeatures.
- Improved material setup and volume parameter synchronization with OutlineVolumeComponent.

- Optimized Kuwahara shaders for better performance and flexibility.
- Refactored AnisotropicKuwahara shader:
  • Replaced runtime loops with precompiled N-value keyword variants (`N_1`, `N_4`, `N_6`, `N_8`).
  • Switched to half precision where possible.
  • Added texture scale control and compression option during the rendering process.
  • Improved parameter accessibility for easier tuning.
- Enhanced KuwaharaGarrettGunnellRendererFeature:
  • Added profiling samplers for each pass (StructureTensor, Eigen, KuwaharaBase, Composite).
  • Implemented texture scaling options and N-value keyword switching.
  • Streamlined RenderGraph passes for cleaner management and reduced overhead.
- Fixed MyPosterizationPass, MyPalettePass, and SharpenPass issues when using secondary shader passes that copy data from temporary textures to the camera color buffer.

- Cleaned up redundant code, improved readability, and increased consistency across all RenderFeatures.
- Setup and renamed PC Render Assets

+++ Assets/Shaders/Kuwahara Implementations/Tonynogo/Kuwahara_TonynogoSettings.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 [Serializable]
 public class Kuwahara_TonynogoSettings
 {
-    [Range(0f, 10f)]
-    public float radius = 0f;
+    [Range(0, 10)]
+    public int radius = 0;
 
     [Range(0, 1f)]
     public float lerpValue = 0f;
